settings Arima

+++ app/src/main/java/com/sweetmay/advancedcryptoindicators2/model/entity/coin/detailed/PriceChangePercentage24hInCurrency.kt
@@ -1,7 +1,7 @@
 package com.sweetmay.advancedcryptoindicators2.model.entity.coin.detailed
 
 data class PriceChangePercentage24hInCurrency(
-        val btc: Double,
-        val rub: Double,
-        val usd: Double,
+        val btc: Float,
+        val rub: Float,
+        val usd: Float,
 )
